+++ src/libTAS/checkpoint/AltStack.cpp
@@ -34,6 +34,13 @@ namespace libtas {
  */
 static stack_t oss;
 
+/* We must save the last stack frame of the alternate stack in the savestate, so
+ * that we return at the correct location from the alternate stack.
+ */
+#define STACKFRAME_OFFSET 0x700
+#define STACKFRAME_SIZE 0x700
+static uint8_t stack_frame[STACKFRAME_SIZE];
+
 void AltStack::saveStack()
 {
     int ret;
@@ -60,6 +67,16 @@ void AltStack::restoreStack()
     int ret;
     NATIVECALL(ret = sigaltstack(&oss, nullptr));
     MYASSERT(ret == 0)
+}
+
+void AltStack::saveStackFrame()
+{
+    memcpy(stack_frame, ReservedMemory::getAddr(ReservedMemory::STACK_ADDR + ReservedMemory::STACK_SIZE - STACKFRAME_OFFSET), STACKFRAME_SIZE);
+}
+
+void AltStack::restoreStackFrame()
+{
+    memcpy(ReservedMemory::getAddr(ReservedMemory::STACK_ADDR + ReservedMemory::STACK_SIZE - STACKFRAME_OFFSET), stack_frame, STACKFRAME_SIZE);
 }
 
 }