+++ src/libTAS/checkpoint/AltStack.h
@@ -28,6 +28,9 @@ namespace AltStack
     void saveStack();
     void prepareStack();
     void restoreStack();
+
+    void saveStackFrame();
+    void restoreStackFrame();
 };
 }