+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Fix SDL1 keyboard layout
 * Wait for threads to register
 * Detect auto-repeat cases when detectable auto-repeat is not supported by the server
+* Fix crash when loading a savestate on a non-draw frame
 
 ## [1.3.1] - 2018-08-30
 ### Added